# backend/swagger.json
{"openapi":"3.1.0","info":{"title":"📱 Mensagens Cooperativa API","description":"\n## Sistema de Gerenciamento de Campanhas de Mensagens WhatsApp\n\nAPI completa para gerenciamento de contatos acadêmicos e disparo de campanhas de mensagens via WhatsApp.\n\n### 🎯 Principais Funcionalidades\n\n* **Autenticação**: Sistema de login com JWT tokens\n* **Contatos**: CRUD completo de contatos acadêmicos\n* **Campanhas**: Criação e gerenciamento de campanhas de mensagens\n* **Dashboard**: Estatísticas e métricas em tempo real\n* **Webhooks**: Recebimento de callbacks da API Z-API\n\n### 🔐 Autenticação\n\nA maioria dos endpoints requer autenticação via Bearer Token. Para obter um token:\n\n1. Faça login no endpoint `/api/v1/auth/login`\n2. Copie o `access_token` da resposta\n3. Use no header: `Authorization: Bearer {seu_token}`\n\n### 📊 Filtros e Paginação\n\nOs endpoints de listagem suportam filtros dinâmicos via query parameters.\n\n### 🚀 Status dos Envios\n\nAs campanhas são processadas em background via Celery + Redis.\nUse os endpoints de dashboard para acompanhar o progresso.\n\n### 📞 Integração Z-API\n\nSistema integrado com Z-API para envio de mensagens WhatsApp.\nConfigure as credenciais no arquivo `.env`.\n    ","termsOfService":"https://mensagenscooperativa.com/termos","contact":{"name":"Suporte Técnico - Mensagens Cooperativa","url":"https://mensagenscooperativa.com/suporte","email":"suporte@mensagenscooperativa.com"},"license":{"name":"Proprietary License","url":"https://mensagenscooperativa.com/licenca"},"version":"1.0.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/webhooks/":{"post":{"tags":["Webhooks"],"summary":"Webhook","operationId":"webhook_api_v1_webhooks__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contacts/":{"get":{"tags":["Contacts"],"summary":"Listar todos os contatos","description":"Retorna lista completa de contatos cadastrados no sistema.","operationId":"list_contacts_api_v1_contacts__get","responses":{"200":{"description":"Lista de contatos retornada com sucesso","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ContactResponse"},"type":"array","title":"Response List Contacts Api V1 Contacts  Get"},"example":[{"id":1,"name":"Maria Silva Santos","phone":"5581987654321","email":"maria.silva@universidade.edu.br","role":"STUDENT","state":"PE","city":"Recife","campus":"Campus Recife","course":"Ciência da Computação","create_at":"2026-01-14T10:30:00"},{"id":2,"name":"João Pedro Oliveira","phone":"5511987654321","email":"joao.pedro@universidade.edu.br","role":"PROFESSOR","state":"SP","city":"São Paulo","campus":"Campus São Paulo","course":"Engenharia de Software","create_at":"2026-01-14T11:00:00"}]}}},"401":{"description":"Não autenticado"}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Contacts"],"summary":"Criar novo contato","description":"Cria um novo contato acadêmico no banco de dados com validação de dados.","operationId":"create_contact_api_v1_contacts__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactCreate"}}},"required":true},"responses":{"201":{"description":"Contato criado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactResponse"},"example":{"id":1,"name":"Maria Silva Santos","phone":"5581987654321","email":"maria.silva@universidade.edu.br","role":"STUDENT","state":"PE","city":"Recife","campus":"Campus Recife","course":"Ciência da Computação","create_at":"2026-01-14T10:30:00"}}}},"400":{"description":"Dados inválidos","content":{"application/json":{"example":{"detail":"Telefone deve conter apenas números"}}}},"401":{"description":"Não autenticado","content":{"application/json":{"example":{"detail":"Token inválido ou expirado"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/contacts/{contact_id}":{"put":{"tags":["Contacts"],"summary":"Atualizar contato existente","description":"Atualiza os dados de um contato específico pelo ID.","operationId":"update_contact_api_v1_contacts__contact_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"contact_id","in":"path","required":true,"schema":{"type":"integer","title":"Contact Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactCreate"}}}},"responses":{"200":{"description":"Contato atualizado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactResponse"},"example":{"id":1,"name":"Maria Silva Santos (Atualizado)","phone":"5581987654321","email":"maria.nova@universidade.edu.br","role":"PROFESSOR","state":"PE","city":"Recife","campus":"Campus Recife","course":"Ciência da Computação","create_at":"2026-01-14T10:30:00"}}}},"404":{"description":"Contato não encontrado","content":{"application/json":{"example":{"detail":"Contato com ID 999 não encontrado"}}}},"401":{"description":"Não autenticado"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Contacts"],"summary":"Deletar contato","description":"Remove permanentemente um contato do sistema pelo ID.","operationId":"delete_contact_api_v1_contacts__contact_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"contact_id","in":"path","required":true,"schema":{"type":"integer","title":"Contact Id"}}],"responses":{"200":{"description":"Contato deletado com sucesso","content":{"application/json":{"schema":{},"example":{"message":"Contato deletado com sucesso"}}}},"404":{"description":"Contato não encontrado","content":{"application/json":{"example":{"detail":"Contato com ID 999 não encontrado"}}}},"409":{"description":"Contato possui campanhas associadas","content":{"application/json":{"example":{"detail":"Não é possível deletar contato com campanhas ativas"}}}},"401":{"description":"Não autenticado"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/campaigns/send":{"post":{"tags":["Campaigns"],"summary":"Criar e enviar campanha","description":"Cria uma nova campanha e inicia o processo de envio em background via Celery.","operationId":"create_and_send_campaign_api_v1_campaigns_send_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignCreate"}}},"required":true},"responses":{"201":{"description":"Campanha criada com sucesso e envio iniciado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponse"},"example":{"id":1,"name":"Divulgação Mestrado 2026","message_body":"Olá! Estamos com inscrições abertas para o Mestrado em Ciência da Computação. Inscreva-se até 28/02/2026!","status":"PENDING","filters_snapshot":{"state":"PE","role":"STUDENT"}}}}},"400":{"description":"Dados inválidos","content":{"application/json":{"example":{"detail":"O corpo da mensagem não pode estar vazio"}}}},"401":{"description":"Não autenticado"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/campaigns/{campaign_id}/messages":{"get":{"tags":["Campaigns"],"summary":"Listar mensagens da campanha","description":"Retorna todas as mensagens enviadas em uma campanha específica com seus status.","operationId":"get_campaign_messages_api_v1_campaigns__campaign_id__messages_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"campaign_id","in":"path","required":true,"schema":{"type":"integer","title":"Campaign Id"}}],"responses":{"200":{"description":"Lista de mensagens retornada com sucesso","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"},"title":"Response Get Campaign Messages Api V1 Campaigns  Campaign Id  Messages Get"},"example":[{"id":1,"contact_id":42,"status":"SENT","zapi_message_id":"3EB0B14F4E7C3E6B5A1F"},{"id":2,"contact_id":43,"status":"DELIVERED","zapi_message_id":"3EB0B14F4E7C3E6B5A2G"},{"id":3,"contact_id":44,"status":"READ","zapi_message_id":"3EB0B14F4E7C3E6B5A3H"},{"id":4,"contact_id":45,"status":"FAILED"}]}}},"404":{"description":"Campanha não encontrada","content":{"application/json":{"example":{"detail":"Campanha com ID 999 não encontrada"}}}},"401":{"description":"Não autenticado"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/":{"get":{"tags":["Dashboard"],"summary":"Get Dashboard","operationId":"get_dashboard_api_v1_dashboard__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"tags":["System"],"summary":"Health Check","description":"Endpoint de verificação de saúde da API. Retorna status e informações básicas do sistema.","operationId":"root__get","responses":{"200":{"description":"API está funcionando corretamente","content":{"application/json":{"schema":{},"example":{"status":"online","message":"Mensagens Cooperativa API is running","version":"1.0.0","database":"mensagens_db","docs":"/docs"}}}}}}}},"components":{"schemas":{"CampaignCreate":{"properties":{"name":{"type":"string","title":"Name"},"message_body":{"type":"string","title":"Message Body"},"scheduled_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Scheduled At"},"filters_snapshot":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Filters Snapshot"}},"type":"object","required":["name","message_body"],"title":"CampaignCreate"},"CampaignResponse":{"properties":{"name":{"type":"string","title":"Name"},"message_body":{"type":"string","title":"Message Body"},"scheduled_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Scheduled At"},"filters_snapshot":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Filters Snapshot"},"id":{"type":"integer","title":"Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["name","message_body","id","status"],"title":"CampaignResponse"},"ContactCreate":{"properties":{"name":{"type":"string","title":"Name"},"phone":{"type":"string","title":"Phone"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"STUDENT"},"campus_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Campus Id"},"academic_area_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Academic Area Id"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"campus":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Campus"},"course":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Course"}},"type":"object","required":["name","phone"],"title":"ContactCreate"},"ContactResponse":{"properties":{"name":{"type":"string","title":"Name"},"phone":{"type":"string","title":"Phone"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"STUDENT"},"campus_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Campus Id"},"academic_area_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Academic Area Id"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"campus":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Campus"},"course":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Course"},"id":{"type":"integer","title":"Id"},"create_at":{"type":"string","format":"date-time","title":"Create At"}},"type":"object","required":["name","phone","id","create_at"],"title":"ContactResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"contact_id":{"type":"integer","title":"Contact Id"},"status":{"type":"string","title":"Status"},"zapi_message_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zapi Message Id"}},"type":"object","required":["id","contact_id","status"],"title":"MessageResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}},"tags":[{"name":"Auth","description":"🔐 Endpoints de autenticação e gerenciamento de sessão. Use o endpoint de login para obter um Bearer token."},{"name":"Contacts","description":"👥 Gerenciamento completo de contatos acadêmicos. CRUD completo com suporte a filtros avançados por estado, cidade, campus e curso."},{"name":"Campaigns","description":"📤 Criação e monitoramento de campanhas de mensagens. As mensagens são enviadas em background via Celery."},{"name":"Dashboard","description":"📊 Estatísticas e métricas do sistema. Visualize total de contatos, campanhas ativas e mensagens enviadas."},{"name":"Webhooks","description":"🔗 Endpoints públicos para receber callbacks da Z-API. Processam status de envio das mensagens."}]}